refactor: lib.rs
