+++ src-tauri/src/types.rs
@@ -1,0 +1,8 @@
+/// Serializable view of the current node.
+#[derive(serde::Serialize)]
+pub struct NodeView {
+    pub id: String,
+    pub text: String,
+    pub end: bool,
+}
+
